cp: polish code
