+++ providers/component-protocol/protocol/framework.go
@@ -70,7 +70,6 @@ func (F FRAMEWORK) Render(ctx context.Context, c *cptype.Component, scenario cpt
 		F.IC.AfterHandleOp(sdk)
 		// encoder
 		ensureCompFieldsBeforeEncode(sdk)
-		//spew.Dump("name:", c.Name, "data ptr", stdStructuredCompPtr.DataPtr())
 		F.IC.EncodeData(stdStructuredCompPtr.DataPtr(), &sdk.Comp.Data)
 		F.IC.EncodeState(stdStructuredCompPtr.StatePtr(), &sdk.Comp.State)
 		F.IC.EncodeInParams(stdStructuredCompPtr.InParamsPtr(), &sdk.InParams)
