+++ providers/component-protocol/protocol/render_parallel.go
@@ -21,7 +21,6 @@ import (
 	"io"
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/sirupsen/logrus"
 
@@ -29,6 +28,7 @@ import (
 	"github.com/erda-project/erda-infra/providers/component-protocol/utils/cputil"
 )
 
+// Node used for parallel rendering.
 type Node struct {
 	Name     string
 	Parallel bool
@@ -274,9 +274,6 @@ func printRenderableNodes(nodes []*Node) {
 	}
 }
 
-//func renderNextNodes(node *Node) error {
-//}
-
 func (n *Node) calcRenderableNextNodes() []*Node {
 	if n.doneNextNodesByName == nil {
 		n.doneNextNodesByName = make(map[string]*Node)
@@ -316,40 +313,4 @@ func (n *Node) calcRenderableNextNodes() []*Node {
 
 func renderOneNode(ctx context.Context, req *cptype.ComponentProtocolRequest, sr ScenarioRender, node *Node) error {
 	return renderOneComp(ctx, req, sr, node.toRendingItem())
-}
-
-func renderOneComp(ctx context.Context, req *cptype.ComponentProtocolRequest, sr ScenarioRender, v cptype.RendingItem) error {
-	// 组件状态渲染
-	err := protoCompStateRending(ctx, req.Protocol, v)
-	if err != nil {
-		logrus.Errorf("protocol component state rending failed, request: %+v, err: %v", v, err)
-		return err
-	}
-	// 获取协议中相关组件
-	c, err := getProtoComp(ctx, req.Protocol, v.Name)
-	if err != nil {
-		logrus.Errorf("get component from protocol failed, scenario: %s, component: %s", req.Scenario.ScenarioKey, req.Event.Component)
-		return nil
-	}
-	// 获取组件渲染函数
-	cr, err := getCompRender(ctx, sr, v.Name, c.Type)
-	if err != nil {
-		logrus.Errorf("get component render failed, scenario: %s, component: %s", req.Scenario.ScenarioKey, req.Event.Component)
-		return err
-	}
-	// 生成组件对应事件，如果不是组件自身事件则为默认事件
-	event := eventConvert(v.Name, req.Event)
-	// 运行组件渲染函数
-	start := time.Now() // 获取当前时间
-	_, instanceName := getCompNameAndInstanceName(v.Name)
-	c.Name = instanceName
-	err = wrapCompRender(cr.RenderC(), req.Protocol.Version).Render(ctx, c, req.Scenario, event, req.Protocol.GlobalState)
-	if err != nil {
-		logrus.Errorf("render component failed, err: %s, scenario: %+v, component: %s", err.Error(), req.Scenario, cr.CompName)
-		return err
-	}
-	simplifyComp(c)
-	elapsed := time.Since(start)
-	logrus.Infof("[component render time cost] scenario: %s, component: %s, cost: %s", req.Scenario.ScenarioKey, v.Name, elapsed)
-	return nil
 }
